requirements.txt
aws-cdk-lib==2.17.0
constructs>=10.0.0,<11.0.0
setuptools==56.0.0
# aws_cdk
awscli==1.22.82
json